Split ExitState.ReturnCheckResults() into helpers
Satisfy gocognit linter, but also (hopefully) make the code
easier to maintain going forward.

refs GH-127
